add primitives for saving color values; make initial primitives depend on `max_node_color`
I realized some time ago that without the ability to save color values
in a stack as we do with positions, there are patterns that are not
easily expressive.  I added this feature and made some minor improvements
to other operations.  I also made the number of integers present in
the initial DSL dependent on `max_node_color`, so that the initial DSL
would always at least contain one integer primitive for every node color.