return type-safe evaluator
The type save evaluator executes the program safely by using GADTS
to structure a miniature interpreter.  Programs are wrapped in a
side-effectful operation that saves the resulting state to the filesystem
so that it can be loaded.  This saving and loading currently works in a
single-threaded manner, but could be generalized.