remove v1 API; implement graph equality based on VF2++ isomorphism testing algorithm
To support procedural generation of graphs, it is crucial to be able to uniquely
identify each node in a neighborhood.  Without this labeling, expansion or traversal
within these neighborhoods is ill-defined.  Previously, I had reused these unique
identifiers of nodes and edges (referred to as "colors" as in colored graphs) to
uniquely identify entire graphs by performing a depth-first traversal starting at
the root node and hashing an encoding of the node colors encountered in this depth-
first traversal.  In other words, I was uniquely identifying graphs quite specifically
with the effectful elements of the programs that generated them.  This meant that
all members of every equivalence class over graph structures was eligible to be
saved and provided as separate possible intermediate representations to downstream
PSNs.  This created a huge amount of rather trivial duplication amongst these discrete
representations.  However, in investigating how graphs could be hashed or compared
for isomorphism more thoroughly, I discovered the research indicating that message-
passing based GNNs are only as powerful as a rather limited form of graph hashing
in detecting graph (non-)isomorphism _unless_ their nodes happen to be uniquely-
labeled.  In the research literature, since possibly isomorphic graphs come from
the wild, there is no unique labeling of nodes that appears particularly sensible.
This resembles the even thornier computational challenge of graph canonization--
the creation of unique representations for graph structures--for which quite limited
solutions are available.  However, in our case, a canonical or preferred form of
the graphs we generate is available: that produced by the most parsimonious
program whose output belongs to a given equivalence class.  By retaining the same
information that uniquely identifies all nodes in all neighborhoods of the graph
for the purposes of procedural generation--but only one such instantiation of this
information--we can create a frontier of programs whose structures are all easily
distinguished by the most pedestrian of message passing-based GNNs, but which do
not feature multiple entries for any one equivalence class.  While there are, of
course, exponentially-many ways to generate any one graph from any of these classes,
invented primitives should bias our program search towards particular ways of
generating those the member of such classes that are reachable within certain search
times or with a particular DSLs.  In inventing these primitives, the compression
mechanism tacitly relies on alignment between graph structure and the procedures
that are used to identify them.  The larger the diversity of graphs from a particular
equivalence class, the larger the diversity of procedures to generate them--
and the less any invention compressing these procedures is responsible for
producing the structures preferred and selected by the PSN.  With this commit,
I hope to lay the groundwork for improving this signal between these elements of
the larger system.

To address this issue, two big changes were necessary:
 - define graph equality based on isomorphism and retain only unique discrete
   representations
 - modify "coloration" to reflect its true purpose of uniquely identifying the
   neighbors of a node; consequently, it is a quality only of a node's edges
   rather than of nodes themselves.

Since the second change was breaking for the v1 API and thus removes the possibility
of using old v1 program synthesis results (which were the only motivation for backwards
compatibility), the v1 API was also deprecated and removed.  I decided to implement the
VF2++ graph isomorphism algorithm, which makes such heavy use of neighbor sets that
it was also worthwhile to define the `Graph.Frozen` module providing an adjacency-set
representation which cannot be updated using functions of the `Operations` module.
This decision interfaces with the design of `Antireduce.Exploration.explore_multikey`
which expects both a primary and a secondary key.  The primary key, in this case,
is given by the histogram of degrees present in a graph, which offers a much lighterweight
way of detecting non-isomorphism.  This histogram is available after a graph has been
converted to a `Graph.Frozen.t`.  The secondary key is the `Graph.Frozen.t` instance,
which may be compared to other instances of this type for isomorphism via
`Graph.Morphism.has_iso`.  The design of `Antireduce.Exploration.multikey_explore`
is tailored specifically to this use case in which some degree of hashing/comparison-
based detection of _in_equality is available for a type, but full equality can only
be tested via a boolean binary equality operation.