remove bidirectionality of edges
I realized that when every edge can be traversed in both directions,
there is very little incentive to learn to use the `add_conn` operation,
as you can always very easily return across paths without planning ahead
as to how to treat graph construction in functional units.  Without this
bidirectionality, a program must plan how to connect new edges back to any
structures it wants to access further.  This should allow programs that
are structurally-sound and aware to have a decisive length advantage
over programs that fail to account for structure.