modify semantics of `add_nb` and remove `func`
`func` is just a composition of `pos_func` and `port_func`, so it should
be learnable on its own.  `add_nb` should not do the work, for the program,
of moving the cursor back to the location from which it was called as
this obviates an otherwise necessary use of `pos_func`--with we want programs
to learn to include on their own.