redesign initial dsl
A persistent problem with the programs synthesized with the previous
initial DSL was the persistent use of `add_nb` over `add_conn`.  I
believe this was because `add_conn` never got off the ground in usage
due to being relatively high-cost.  To use it, one had to traverse to
two locations that one had already visited from a third location.
This made it always cheaper to use `add_nb` to simply add a neighbor.
Since no more than this was every required to produce unique graphs,
it would make sense that the more complex usage of `add_conn` would
never arise.  The fundamental issue was that there were two connection
operations, one of which was also the only operation that expanded the
graph.

In the redesign, a stack of positions is used to support addition of
neighboring edges and all edges are made bidirectional.  Additionally,
for simplicity only one function supporting subroutines (`func`) is
present rather than `pos_func` and `port_func`, which restores the state
of the stack, position, and port.  Adding a node is now a separate
operation (`add`), and connection always occurs between the node at
the top of the stack and the current position.  The only complication of
this design is that if we want to ensure that graphs consist of a single
connected component, we must require that every node that is added is
eventually connected or the resulting graph is discarded.  The simplest
way to do this is to track the number of connected components in a graph,
incrementing this number when a new node is added, and not to allow the
addition of a new isolated node until this number has been decreased
back to 1.  As long as `add` relocates the cursor to the newly-created
node, this ensures that no further traversal or modification of the
graph can take place until an edge is added.  Some further optimizations
have been added as well, such as a check that there is at least one node
on the stack before `add` is called, and that `pop` cannot remove the
last node necessary to connect an isolated node.  These ensure that it is
always possible to recover from an `add` operation.  Nonetheless, if
a graph program does not do so, it is simple to discard it as the
number of connected components is an attribute of the record `Graph.t`
rather than `State.t`.