fix bug in `func` initial DSL primitive implementation
The implementation of this primitive has to handle the presence
of an isolated node.  I think the best way to do this without
rewarding the creation of isolated nodes is to only allow the
transformation performed by the first argument of `func` to take
effect if the graph it produces has a single connected component.